+++ client/src/components/new game/step 1/EmailInvite.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { useEmails } from "../../../DataContext";
+import { useEmails } from "../../../contexts/DataContext";
 import { useState } from "react";
 import {
   Typography,
